Move LogCode enum to domain module
