+++ application/usecase/replace.py
@@ -12,7 +12,7 @@ from ...domain.port.state import StateRepository
 from ...domain.service.history import policy as history_policy
 from ...domain.value.content import Payload, resolve_content
 from ...domain.value.message import Scope
-from ...logging.code import LogCode
+from ...domain.log.code import LogCode
 
 logger = logging.getLogger(__name__)
