+++ presentation/telegram/router.py
@@ -10,7 +10,7 @@ from ..navigator import Navigator
 from ...application.log.emit import jlog
 from ...domain.error import HistoryEmpty, InlineUnsupported
 from ...domain.value.lexicon import lexeme
-from ...logging.code import LogCode
+from ...domain.log.code import LogCode
 
 router = Router(name="navigator_handlers")
